compile: Move 'exprtree' package up into new 'backend'
This doesn't introduce any change in logic.
