+++ src/main/java/com/abusalimov/mrcalc/backend/PrimitiveCastBuilder.java
@@ -1,4 +1,4 @@
-package com.abusalimov.mrcalc.compile.exprtree;
+package com.abusalimov.mrcalc.backend;
 
 /**
  * @author Eldar Abusalimov
